refactor(zhenxun): 调整 unicode_escape 和 unicode_unescape 函数位置
- 从 db_context.py 中移除这两个函数
- 将它们添加到 utils.py 中，以更好地组织代码
